+++ Demo_MondrianLayout/Book.Overlay.swift
@@ -8,7 +8,7 @@ var _book_overlay: BookView {
     BookPreview {
       ExampleView(width: nil, height: nil) { (view: UIView) in
 
-        view.mondrian.buildSubviews {
+        Mondrian.buildSubviews(on: view) {
           UIView.mock(
             backgroundColor: .mondrianYellow,
             preferredSize: .init(width: 100, height: 100)
@@ -31,7 +31,7 @@ var _book_overlay: BookView {
 
     BookPreview {
       ExampleView(width: nil, height: nil) { (view: UIView) in
-        view.mondrian.buildSubviews {
+        Mondrian.buildSubviews(on: view) {
           UIView.mock(
             backgroundColor: .mondrianYellow,
             preferredSize: .init(width: 100, height: 100)
@@ -53,7 +53,7 @@ var _book_overlay: BookView {
 
     BookPreview {
       ExampleView(width: nil, height: nil) { (view: UIView) in
-        view.mondrian.buildSubviews {
+        Mondrian.buildSubviews(on: view) {
           VStackBlock {
             UIView.mock(
               backgroundColor: .mondrianYellow,
@@ -76,7 +76,7 @@ var _book_overlay: BookView {
 
     BookPreview {
       ExampleView(width: nil, height: nil) { (view: UIView) in
-        view.mondrian.buildSubviews {
+        Mondrian.buildSubviews(on: view) {
           VStackBlock(spacing: 2) {
             UIView.mock(
               backgroundColor: .mondrianYellow,
@@ -115,7 +115,7 @@ var _book_overlay: BookView {
 
     BookPreview {
       ExampleView(width: nil, height: nil) { (view: UIView) in
-        view.mondrian.buildSubviews {
+        Mondrian.buildSubviews(on: view) {
           VStackBlock {
             UIView.mock(
               backgroundColor: .mondrianYellow,
