+++ Tests/VGridTests.swift
@@ -15,7 +15,7 @@ final class VGridTests: XCTestCase {
   func test_basic() {
 
     let view = ExampleView(width: 100, height: nil) { view in
-      view.mondrian.buildSubviews {
+      Mondrian.buildSubviews(on: view) {
         VGridBlock(
           columns: [
             .init(.flexible(), spacing: 16),
